requirements.txt
# Core IR Libraries
bm25s>=0.2.0,<0.4.0
sentence-transformers>=2.2.0,<3.0.0
ragatouille>=0.0.8

# Optimization & Acceleration
onnxruntime>=1.15.0
optimum>=1.14.0

# NLP & Preprocessing
spacy>=3.7.0,<4.0.0
transformers>=4.35.0,<5.0.0
torch>=2.0.0

# Topic Modeling
bertopic>=0.15.0
umap-learn>=0.5.0

# Data & Storage
pandas>=2.0.0
numpy>=1.24.0,<2.0.0

# Evaluation & Visualization
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.18.0

# User Interface
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
nltk>=3.8.0

pyspellchecker
rank_bm25